backend/requirements.txt
django==5.1
djangorestframework==3.15.2
pytest-django==4.9.0
model-bakery==1.19.5
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1
Pillow==11.0.0